frontend/requirements.txt

streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
openpyxl>=3.1.0
python-dotenv
pytz
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
